Some refactoring and cleanup.

+++ 2025/Day01/Day01.h
@@ -3,8 +3,8 @@
 #include <string>
 
 namespace Day01 {
-    float firstPart();
-    float secondPart();
+    double firstPart();
+    double secondPart();
 
     inline std::string inputPath = "C:\\Users\\lenna\\Documents\\Workspace\\Programming\\AdventOfCode\\2025\\Day01\\input.txt";
     inline std::string examplePath = "C:\\Users\\lenna\\Documents\\Workspace\\Programming\\AdventOfCode\\2025\\Day01\\input.txt";
