+++ 2025/Day02/Day02.h
@@ -2,8 +2,8 @@
 #include <string>
 
 namespace Day02 {
-    float firstPart();
-    float secondPart();
+    double firstPart();
+    double secondPart();
 
     inline std::string inputPath = "C:\\Users\\lenna\\Documents\\Workspace\\Programming\\AdventOfCode\\2025\\Day02\\input.txt";
     inline std::string examplePath = "C:\\Users\\lenna\\Documents\\Workspace\\Programming\\AdventOfCode\\2025\\Day02\\example.txt";
