+++ 2025/Day03/Day03.cpp
@@ -6,14 +6,14 @@
 
 using namespace Day03;
 
-float Day03::firstPart() {
+double Day03::firstPart() {
     auto input = Utility::readInput(inputPath);
     int result = -1;
 
     return result;
 }
 
-float Day03::secondPart() {
+double Day03::secondPart() {
     auto input = Utility::readInput(inputPath);
     int result = -1;
